7주차 > advanced > 전역 integerated hook 작성, app wrapping, 컴포넌트 인터페이스 일부 정리
