+++ packages/assignment-7/src/main.tsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
+import Scheduler from "./Scheduler/index.tsx";
 
 async function enableMocking() {
   const { worker } = await import("./mocks/mockServiceWorker/browser.ts");
@@ -13,7 +13,7 @@ enableMocking().then(() => {
   ReactDOM.createRoot(document.getElementById("root")!).render(
     <React.StrictMode>
       <ChakraProvider>
-        <App />
+        <Scheduler />
       </ChakraProvider>
     </React.StrictMode>
   );
